refactor(verifier): consolidate public types
