+++ packages/core/lib/v3/verifier/types.ts
@@ -1,0 +1,339 @@
+/**
+ * Shared verifier types for trajectories, rubrics, evidence, and verdicts.
+ *
+ * The verifier consumes saved trajectories instead of a live browser. DOM and
+ * Hybrid runs preserve tool-return text/JSON evidence, while CUA runs preserve
+ * screenshots sent to the provider plus independent harness probes.
+ */
+
+/** Token usage for one or more LLM calls. Matches AgentResult.usage shape. */
+export interface TrajectoryUsage {
+  input_tokens: number;
+  output_tokens: number;
+  reasoning_tokens?: number;
+  cached_input_tokens?: number;
+  inference_time_ms?: number;
+}
+
+/**
+ * A single criterion in a Stagehand rubric. Dataset and model wire formats may
+ * use serialized `max_points` / `earned_points`; normalize those with
+ * `normalizeRubric()` at the boundary.
+ */
+export interface RubricCriterion {
+  /** Short name of the criterion (e.g., "Add ground beef to cart"). */
+  criterion: string;
+  /** What to evaluate and how to award partial credit. */
+  description: string;
+  /** Maximum points for this criterion. */
+  maxPoints: number;
+  /**
+   * Triggering condition for conditional criteria. Only counted when met
+   * (paper's "Mutually Exclusive Conditionals" pattern).
+   */
+  condition?: string;
+  /** Filled by the verifier during scoring; empty in precomputed rubrics. */
+  justification?: string;
+  /**
+   * Filled by the verifier during scoring; empty string in some serialized
+   * upstream rubrics and a number in scored rubrics.
+   */
+  earnedPoints?: number | string;
+}
+
+/** A rubric — list of criteria for a task. */
+export interface Rubric {
+  items: RubricCriterion[];
+}
+
+/**
+ * Serialized rubric item shape as stored in datasets and prompt responses.
+ * Keep this at IO boundaries; core verifier types use camelCase.
+ */
+export interface SerializedRubricCriterion {
+  criterion: string;
+  description: string;
+  max_points: number;
+  condition?: string;
+  justification?: string;
+  earned_points?: number | string;
+}
+
+/** Serialized rubric shape used by upstream datasets and generated JSON. */
+export interface SerializedRubric {
+  items: SerializedRubricCriterion[];
+}
+
+export type RubricInput = Rubric | SerializedRubric;
+
+/**
+ * Spec for a single task being verified. Carried both at runtime and into the
+ * verifier alongside the trajectory.
+ */
+export interface TaskSpec {
+  /** Stable identifier (e.g., "united_13" for WebTailBench, task_id for Mind2Web). */
+  id: string;
+  /** Task instruction shown to the agent. */
+  instruction: string;
+  /** Starting URL, if any. */
+  initUrl?: string;
+  /** Rubric carried by the dataset or generated by a verifier backend. */
+  precomputedRubric?: Rubric;
+  /** Optional reference answer (set when dataset ships one). */
+  expectedAnswer?: string;
+}
+
+/**
+ * A single modality unit in tier-1 agent evidence. Mirrors the shape of
+ * ModelMessage content parts so we can reproduce what the LLM ingested.
+ */
+export type AgentEvidenceModality =
+  | { type: "text"; content: string }
+  | { type: "image"; bytes: Buffer; mediaType: string }
+  | { type: "json"; content: unknown };
+
+/**
+ * Tier 1 — exactly the bytes/strings/objects the agent's LLM ingested as the
+ * tool result for this step.
+ *
+ * Modes:
+ *   - CUA: usually a single image modality (the screenshot sent to the provider).
+ *   - Hybrid: tool result with optional screenshotBase64 → one image + one text.
+ *   - DOM: tool returns (extract JSON, ariaTree text, etc.) → text/json modalities.
+ */
+export interface AgentEvidence {
+  modalities: AgentEvidenceModality[];
+}
+
+/**
+ * Tier 2 — independent harness probes around this step.
+ *
+ * If a probe wasn't captured, the field is absent (not null).
+ */
+export interface ProbeEvidence {
+  /** URL after the step's tool execution. */
+  url?: string;
+  /**
+   * Bus screenshot captured after the step. Path on disk is preferred once
+   * persisted; in-memory Buffer is used during a live run.
+   */
+  screenshot?: Buffer;
+  /** Reference to the persisted screenshot file under the trajectory dir. */
+  screenshotPath?: string;
+  /** Viewport scroll context. Lets the verifier reason about whether the agent saw the full page. */
+  scroll?: { top: number; pageHeight: number };
+  /** Accessibility tree snapshot. */
+  ariaTree?: string;
+  /** Verifier-requested probes, keyed by criterion id. */
+  onDemand?: Record<string, unknown>;
+}
+
+/** Outcome of a single tool execution as seen by the harness. */
+export interface ToolOutput {
+  ok: boolean;
+  /**
+   * The tool's return value. Same payload that flowed into agentEvidence
+   * modalities, but in its native shape (e.g., the extract result, the act
+   * describe-string) rather than serialized for the LLM.
+   */
+  result: unknown;
+  error?: string;
+}
+
+/** One step in a trajectory: action + reasoning + evidence + outcome. */
+export interface TrajectoryStep {
+  index: number;
+  actionName: string;
+  actionArgs: Record<string, unknown>;
+  /** From AgentAction.reasoning. May be empty for tools that don't surface reasoning. */
+  reasoning: string;
+  agentEvidence: AgentEvidence;
+  probeEvidence: ProbeEvidence;
+  toolOutput: ToolOutput;
+  /** ISO 8601 timestamp when the step's tool execution started. */
+  startedAt: string;
+  /** ISO 8601 timestamp when the step's tool execution finished. */
+  finishedAt: string;
+}
+
+/** Terminal status of the agent run. */
+export type TrajectoryStatus = "complete" | "aborted" | "stalled" | "error";
+
+/**
+ * Full trajectory for one task run.
+ *
+ * The on-disk layout is one directory per task:
+ *
+ *   .trajectories/<run-id>/<task-id>/
+ *     ├── task_data.json    — TaskSpec + Verdict (filled on completion)
+ *     ├── trajectory.json   — this object, with screenshotPath instead of bytes
+ *     ├── screenshot_1.png  — probeEvidence.screenshot for step 1, etc.
+ *     ├── scores/
+ *     │   └── mmrubric_v1.json  — Verdict from V3Evaluator.verify()
+ *     ├── core.log          — captured action log
+ *     └── times.json        — step timing + token usage
+ */
+export interface Trajectory {
+  task: TaskSpec;
+  steps: TrajectoryStep[];
+  finalAnswer?: string;
+  status: TrajectoryStatus;
+  usage: TrajectoryUsage;
+  timing: { startedAt: string; endedAt: string };
+}
+
+/** Score for a single rubric criterion after evidence analysis + rescoring. */
+export interface CriterionScore {
+  /** Matches RubricCriterion.criterion (the criterion's short name). */
+  criterion: string;
+  /** Maximum possible points for this criterion. */
+  maxPoints: number;
+  /**
+   * Points earned post-evidence-analysis (paper's post_image_earned_points).
+   * Null if the criterion was conditional and its condition wasn't met (excluded
+   * from both numerator and denominator in the process score).
+   */
+  earnedPoints: number | null;
+  /** Verifier's free-text justification for the score. */
+  justification: string;
+  /**
+   * True if the criterion is conditional and its condition was determined to
+   * be met. Absent for non-conditional criteria.
+   */
+  conditionMet?: boolean;
+  /**
+   * Set when the verifier had no evidence to ground this criterion in either
+   * tier. Per paper §2, treated as uncontrollable failure → full credit, but
+   * surfaced here so dashboards can flag low-confidence verdicts.
+   */
+  evidenceInsufficient?: boolean;
+}
+
+/**
+ * First-point-of-failure analysis (paper Step 9a). Identifies the earliest
+ * step where the agent's trajectory went off-track, using a structured error
+ * taxonomy (7 top-level categories, 1.1–7.4 sub-codes).
+ */
+export interface FirstPointOfFailure {
+  stepIndex: number;
+  /** Sub-code from the error taxonomy (e.g., "2.3" for a specific hallucination type). */
+  errorCode: string;
+  /** Top-level category name (Selection, Hallucination, etc.). */
+  category: string;
+  /** Verifier's reasoning for selecting this point. */
+  description?: string;
+}
+
+/**
+ * Structured observation surfaced by the verifier that another agent or
+ * tooling could act on. Findings are emitted opportunistically by Step 8
+ * (outcome verification) when the verifier notices actionable patterns —
+ * repeated tool-call failures, ambiguous task specs, evidence gaps, etc.
+ *
+ * Not produced for every task: when nothing actionable surfaces, the
+ * `findings` array on the Verdict is empty. Consumers should treat the
+ * field as advisory, not as part of the formal score.
+ */
+export interface VerifierFinding {
+  /**
+   * Category of the observation. Open-ended enum — additional categories may
+   * be added as verifier backends surface new failure modes.
+   */
+  category:
+    | "agent_tool_usage" // agent's tool calls had repeated issues (misclicks, wrong args, retries)
+    | "agent_strategy" // higher-level planning / decision-making problems
+    | "rubric_quality" // criteria were overly strict, ambiguous, or contradictory
+    | "trajectory_capture" // gaps in evidence (missing screenshots, empty steps)
+    | "task_specification" // task instruction was ambiguous / under- or over-specified
+    | "verifier_uncertainty" // verifier itself couldn't confidently decide
+    | "other";
+  /** Impact: info (FYI), warning (worth investigating), blocking (broke the task). */
+  severity: "info" | "warning" | "blocking";
+  /** What the verifier noticed. Plain prose, grounded in evidence from the trajectory. */
+  description: string;
+  /**
+   * Optional concrete next action another agent could take. Should be
+   * specific enough that it can be acted on without further reasoning —
+   * e.g., "Try double_click instead of triple_click to clear placeholder
+   * text on this form field."
+   */
+  suggestedAction?: string;
+  /** Step indices in the trajectory where this pattern showed up. */
+  relatedSteps?: number[];
+}
+
+/** Stable debugging summary emitted by verifier backends. */
+export interface VerifierRawSteps {
+  backend?: "legacy" | "verifier";
+  primaryIntent?: string;
+  reasoning?: string;
+  rubricSource?: "precomputed" | "generated" | "none";
+  approach?: "a" | "b";
+  optionalsMode?: "folded" | "separate" | "skip";
+  totalEarned?: number;
+  totalMax?: number;
+  evidenceImages?: number;
+  evidenceTexts?: number;
+  evidenceOriginalScreenshots?: number;
+  legacyEvaluation?: string;
+  screenshotCount?: number;
+}
+
+/** Task-validity classification (paper Step 10). */
+export interface TaskValidity {
+  /** True if the task is underspecified / has multiple valid interpretations. */
+  isAmbiguous: boolean;
+  /** True if the task is impossible / illegal / NSFW / otherwise infeasible. */
+  isInvalid: boolean;
+  /** Optional sub-codes from the task-classification taxonomy. */
+  ambiguityCodes?: string[];
+  invalidTaskCodes?: string[];
+}
+
+/**
+ * The verifier's output. Process score + outcome verdict + diagnostic signals.
+ *
+ * Process and outcome are deliberately independent (paper §2): an agent can
+ * follow the right steps but get blocked (high process, low outcome), or
+ * succeed through an unexpected path (variable process, high outcome).
+ */
+export interface Verdict {
+  /** Step 8 — did the agent accomplish the task from the user's perspective? */
+  outcomeSuccess: boolean;
+  /** Aggregated earned/max across applicable criteria, in [0, 1]. */
+  processScore: number;
+  /** Per-criterion breakdown after rescoring. */
+  perCriterion: CriterionScore[];
+  /** Step 9a — first step where the trajectory went off-track, if any. */
+  firstPointOfFailure?: FirstPointOfFailure;
+  /** Step 10 — task-itself ambiguity / validity. */
+  taskValidity: TaskValidity;
+  /**
+   * Ids (RubricCriterion.criterion strings) of criteria where neither tier of
+   * evidence resolved the question. Treated as uncontrollable → full credit,
+   * but flagged here so consumers can decide whether to discount the score.
+   */
+  evidenceInsufficient: string[];
+  /**
+   * Structured observations from the verifier that a downstream tool or
+   * follow-up agent could act on. Opportunistic — empty when the verifier
+   * doesn't notice anything actionable. Not part of the score; advisory.
+   */
+  findings?: VerifierFinding[];
+  /** Debugging summary from the active evaluator backend. */
+  rawSteps?: VerifierRawSteps;
+}
+
+/** Reason a stub verifier can emit. */
+export type StubVerdictReason =
+  | "wave-0-stub"
+  | "no-rubric"
+  | "empty-trajectory";
+
+/**
+ * Verifier interface. Implementations consume a Trajectory + TaskSpec and
+ * return a Verdict — they MUST NOT touch a live browser.
+ */
+export interface Verifier {
+  verify(trajectory: Trajectory, taskSpec: TaskSpec): Promise<Verdict>;
+}
